requirements.txt
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.172.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
langchain==0.3.25
langchain-community==0.3.25
langchain-core==0.3.65
langchain-google-genai==2.0.10
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.8
langsmith==0.3.45
